+++ packages/create-app/templates/vue2-static/vite.config.js
@@ -1,11 +1,10 @@
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import vue from '@vitejs/plugin-vue2';
 import { createBlockletPlugin } from 'vite-plugin-blocklet';
-import svgr from 'vite-plugin-svgr';
 
 // https://vitejs.dev/config/
 export default defineConfig(() => {
   return {
-    plugins: [react(), createBlockletPlugin(), svgr()],
+    plugins: [vue(), createBlockletPlugin()],
   };
 });