+++ plugins/vite-plugin-blocklet/index.js
@@ -1,6 +1,7 @@
 import createHmrPlugin from './libs/hmr.js';
 import createConfigPlugin from './libs/config.js';
 import createMetaPlugin from './libs/meta.js';
+import setupClient from './libs/client';
 
 export function createBlockletPlugin(options = {}) {
   const { disableConfig = false, disableMeta = false, disableHmr = false } = options;
@@ -18,6 +19,9 @@ export function createBlockletPlugin(options = {}) {
   return plugins;
 }
 
-export const createBlockletHmr = createHmrPlugin;
-export const createBlockletConfig = createConfigPlugin;
-export const createBlockletMeta = createMetaPlugin;
+export {
+  setupClient,
+  createHmrPlugin as createBlockletHmr,
+  createConfigPlugin as createBlockletConfig,
+  createMetaPlugin as createBlockletMeta,
+};